camera_OD/requirements.txt
Pillow
onnx==1.6.0
numpy
matplotlib
opencv-python
